+++ Changelog.md
@@ -12,6 +12,7 @@
 2. Fix: Correct a typo in code display: `/docs/usages/data`.
 3. Fix: Rows rendered by `react-json-grid` in this document should not be styled by the colors of Docusaurus.
 4. Fix: Add a one-time hook to the demo apps to ensure that they will be updated by the site theme if necessary.
+5. Fix: Change the implementation of `ThemedComponent` to ensure that the dark theme can be applied correctly.
 
 #### :floppy_disk: Change